Extract Area into it's own file
