+++ lib/everypolitician/popolo/area.rb
@@ -1,0 +1,18 @@
+module Everypolitician
+  module Popolo
+    class Areas < Collection
+      def initialize(documents)
+        @documents = documents.map { |p| Area.new(p) }
+      end
+    end
+
+    class Area
+      def initialize(document)
+        @document = document
+        document.each do |key, value|
+          define_singleton_method(key) { value }
+        end
+      end
+    end
+  end
+end
